+++ Sources/Statistics/Statistics.swift
@@ -68,6 +68,7 @@ func mean(n: Int, p: Int, q: Int) -> Double {
 
 public extension Collection where Iterator.Element == Float {
     func mean() -> Float {
+        guard count > 0 else { return 0 }
         return sum() / Float(count)
     }
     
@@ -90,17 +91,29 @@ public extension Collection where Iterator.Element == Float {
         })
     }
     
-    func histogram(buckets: [Range<Float>]) -> [Int] {
-        return buckets.map({ (bucket) -> Int in
-            return filter({ (value) -> Bool in
+    func histogram(buckets: [Range<Float>]) -> HistogramCount {
+        return buckets.reduce(into: HistogramCount(), { (result, bucket) in
+            result[String(describing: bucket)] = filter({ (value) -> Bool in
                 return bucket.contains(value)
             }).count
         })
+    }
+    
+    func normalDistribution() -> HistogramCount {
+        let zScores = self.zScores()
+        let maxZ = ceil(zScores.max()!)
+        let minZ = floor(zScores.min()!)
+        let maxDistanceFromMean = Swift.max(abs(maxZ), abs(minZ))
+        let buckets = stride(from: -maxDistanceFromMean, through: maxDistanceFromMean, by: 1).map({ (bucket) -> Range<Float> in
+            return Range(uncheckedBounds: (lower: bucket, upper: bucket + 1))
+        })
+        return zScores.histogram(buckets: buckets)
     }
 }
 
 public extension Collection where Iterator.Element == Double {
     func mean() -> Double {
+        guard count > 0 else { return 0 }
         return sum() / Double(count)
     }
     
@@ -123,25 +136,39 @@ public extension Collection where Iterator.Element == Double {
         })
     }
     
-    func histogram(buckets: [Range<Double>]) -> [Int] {
-        return buckets.map({ (bucket) -> Int in
-            return filter({ (value) -> Bool in
+    func histogram(buckets: [Range<Double>]) -> HistogramCount {
+        return buckets.reduce(into: HistogramCount(), { (result, bucket) in
+            result[String(describing: bucket)] = filter({ (value) -> Bool in
                 return bucket.contains(value)
             }).count
         })
+    }
+    
+    func normalDistribution() -> HistogramCount {
+        let zScores = self.zScores()
+        let maxZ = ceil(zScores.max()!)
+        let minZ = floor(zScores.min()!)
+        let maxDistanceFromMean = Swift.max(abs(maxZ), abs(minZ))
+        let buckets = stride(from: -maxDistanceFromMean, through: maxDistanceFromMean, by: 1).map({ (bucket) -> Range<Double> in
+            return Range(uncheckedBounds: (lower: bucket, upper: bucket + 1))
+        })
+        return zScores.histogram(buckets: buckets)
     }
 }
 
 public extension Collection where Iterator.Element == Int {
     func mean() -> Double {
+        guard count > 0 else { return 0 }
         return Double(sum()) / Double(count)
     }
     
     func variance() -> Double {
         let mean = self.mean()
-        return map({ (next) -> Double in
+        let squaredDifferences = map({ (next) -> Double in
             return pow(Double(next) - mean, 2)
-        }).mean()
+        })
+        let variance = squaredDifferences.mean()
+        return variance
     }
     
     func standardDeviation() -> Double {
@@ -156,11 +183,62 @@ public extension Collection where Iterator.Element == Int {
         })
     }
     
-    func histogram(buckets: [Range<Int>]) -> [Int] {
-        return buckets.map({ (bucket) -> Int in
-            return filter({ (value) -> Bool in
+    func histogram(buckets: [Range<Int>]) -> HistogramCount {
+        return buckets.reduce(into: HistogramCount(), { (result, bucket) in
+            result[String(describing: bucket)] = filter({ (value) -> Bool in
                 return bucket.contains(value)
             }).count
         })
+    }
+    
+    func normalDistribution() -> HistogramCount {
+        let zScores = self.zScores()
+        let maxZ = ceil(zScores.max()!)
+        let minZ = floor(zScores.min()!)
+        let maxDistanceFromMean = Swift.max(abs(maxZ), abs(minZ))
+        let buckets = stride(from: -maxDistanceFromMean, through: maxDistanceFromMean, by: 1).map({ (bucket) -> Range<Double> in
+            return Range(uncheckedBounds: (lower: bucket, upper: bucket + 1))
+        })
+        return zScores.histogram(buckets: buckets)
+    }
+}
+
+/// Map of bucket names to frequencies
+public typealias HistogramCount = [String: Int]
+
+/// Map of bucket names to tuples of frequencies and the set of tags that are included
+public typealias TagValueMap = [String: Double]
+public typealias TaggedCount = (count: Int, tags: Set<String>)
+public typealias TaggedHistogramCount = [Double: TaggedCount]
+
+public extension Dictionary where Key == String, Value == Double {
+    func zScores() -> TagValueMap {
+        let values = self.map({$0.value})
+        let mean = values.mean()
+        let standardDeviation = values.standardDeviation()
+        return reduce(into: [String: Double](), { (result, next) in
+            let (key, value) = next
+            result[key] = (value - mean) / standardDeviation
+        })
+    }
+    
+    func histogram(buckets: [Range<Double>]) -> TaggedHistogramCount {
+        return buckets.reduce(into: TaggedHistogramCount(), { (result, bucket) in
+            let items = filter({ (next) -> Bool in
+                return bucket.contains(next.value)
+            })
+            result[bucket.upperBound] = (items.count, Set(items.keys))
+        })
+    }
+    
+    func normalDistribution() -> TaggedHistogramCount {
+        let zScores = self.zScores()
+        let maxZ = ceil(zScores.values.max()!)
+        let minZ = floor(zScores.values.min()!)
+        let maxDistanceFromMean = Swift.max(abs(maxZ), abs(minZ))
+        let buckets = stride(from: -maxDistanceFromMean, through: maxDistanceFromMean, by: 1).map({ (bucket) -> Range<Double> in
+            return Range(uncheckedBounds: (lower: bucket, upper: bucket + 1))
+        })
+        return zScores.histogram(buckets: buckets)
     }
 }